+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ScenarioActivityApiSpec.scala
@@ -10,7 +10,7 @@ import pl.touk.nussknacker.ui.api.helpers.{NuItTest, NuTestScenarioManager, With
 
 import java.util.UUID
 
-class ScenarioActivitySpec
+class ScenarioActivityApiSpec
     extends AnyFreeSpecLike
     with NuItTest
     with WithMockableDeploymentManager
@@ -216,8 +216,8 @@ class ScenarioActivitySpec
           .applicationState {
             createSavedScenario(exampleScenario)
           }
-          .multiPartBody(fileContent = fileContent, fileName = fileName)
-          .basicAuth("writer", "writer")
+          .streamBody(fileContent = fileContent, fileName = fileName)
+          .preemptiveBasicAuth("writer", "writer")
           .when()
           .post(s"$nuDesignerHttpAddress/api/processes/$exampleScenarioName/1/activity/attachments")
           .Then()
@@ -234,7 +234,7 @@ class ScenarioActivitySpec
             createAttachment(scenarioName = exampleScenarioName, fileContent = fileContent1, fileName = fileName)
             createAttachment(scenarioName = exampleScenarioName, fileContent = fileContent2, fileName = fileName)
           }
-          .basicAuth("reader", "reader")
+          .preemptiveBasicAuth("reader", "reader")
           .when()
           .get(s"$nuDesignerHttpAddress/api/processes/$exampleScenarioName/activity")
           .Then()
@@ -267,8 +267,8 @@ class ScenarioActivitySpec
 
       "return 404 for no existing scenario" in {
         given()
-          .multiPartBody(fileContent = fileContent, fileName = fileName)
-          .basicAuth("writer", "writer")
+          .streamBody(fileContent = fileContent, fileName = fileName)
+          .preemptiveBasicAuth("writer", "writer")
           .when()
           .post(s"$nuDesignerHttpAddress/api/processes/$wrongScenarioName/1/activity/attachments")
           .Then()
@@ -280,7 +280,7 @@ class ScenarioActivitySpec
     "not authenticated should" - {
       "forbid access for no authorization" in {
         given()
-          .multiPartBody(fileContent = "test", fileName = "test.xml")
+          .streamBody(fileContent = "test", fileName = "test.xml")
           .noAuth()
           .when()
           .post(s"$nuDesignerHttpAddress/api/processes/$exampleScenarioName/1/activity/attachments")
@@ -294,8 +294,8 @@ class ScenarioActivitySpec
           .applicationState {
             createSavedScenario(exampleScenario)
           }
-          .multiPartBody(fileContent = "test", fileName = "test.xml")
-          .basicAuth("reader", "reader")
+          .streamBody(fileContent = "test", fileName = "test.xml")
+          .preemptiveBasicAuth("reader", "reader")
           .when()
           .post(s"$nuDesignerHttpAddress/api/processes/$exampleScenarioName/1/activity/attachments")
           .Then()
@@ -379,8 +379,8 @@ class ScenarioActivitySpec
       fileName: String = "important_file.txt"
   ): Unit = {
     given()
-      .multiPartBody(fileContent = fileContent, fileName = fileName)
-      .basicAuth("writer", "writer")
+      .streamBody(fileContent = fileContent, fileName = fileName)
+      .preemptiveBasicAuth("writer", "writer")
       .when()
       .post(s"$nuDesignerHttpAddress/api/processes/$scenarioName/1/activity/attachments")
   }